feat/gateway: added redis-backed rate limiting bucket

+++ services/gateway/src/main/java/org/bumIntra/gateway/security/ratelimit/RateLimitProfile.java
@@ -2,5 +2,5 @@ package org.bumIntra.gateway.security.ratelimit;
 
 import java.time.Duration;
 
-public record RateLimitProfile(int capacity, Duration refillPeriod) {
+public record RateLimitProfile(long capacity, Duration refillPeriod) {
 }
